Move deprecated methods to standalone file
