+++ LineSDK/LineSDK/Networking/API/API+Deprecated.swift
@@ -1,5 +1,5 @@
 //
-//  API.swift
+//  API+Deprecated.swift
 //
 //  Copyright (c) 2016-present, LINE Corporation. All rights reserved.
 //
@@ -21,18 +21,7 @@
 
 import Foundation
 
-/// Represents a utility structure for calling the LINE Platform.
-///
-/// - Note:
-/// For most API calls, using the interfaces in the `API` structure is equivalent to using and sending an
-/// underlying `Request` object with a `Session` object. However, some methods in the `API` provides
-/// additional useful features such as working with the keychain and redirecting the final result in a more
-/// reasonable way.
-///
-/// Using the `API` structure to interact with the LINE Platform is highly recommended unless you are familiar
-/// with and want to extend the LINE SDK to send unimplemented API requests to the LINE Platform.
-///
-public enum API {
+extension API {
     /// Refreshes the access token with `refreshToken`.
     ///
     /// - Parameters:
@@ -210,38 +199,6 @@ public enum API {
             return
         }
         let request = GetVerifyTokenRequest(accessToken: token)
-        Session.shared.send(request, callbackQueue: queue, completionHandler: completion)
-    }
-    
-    /// Gets the user's profile.
-    ///
-    /// - Parameters:
-    ///   - queue: The callback queue that is used for `completion`. The default value is
-    ///            `.currentMainOrAsync`. For more information, see `CallbackQueue`.
-    ///   - completion: The completion closure to be invoked when the user's profile is returned.
-    /// - Note: The `.profile` permission is required.
-    ///
-    public static func getProfile(
-        callbackQueue queue: CallbackQueue = .currentMainOrAsync,
-        completionHandler completion: @escaping (Result<UserProfile, LineSDKError>) -> Void)
-    {
-        let request = GetUserProfileRequest()
-        Session.shared.send(request, callbackQueue: queue, completionHandler: completion)
-    }
-    
-    /// Gets the friendship status of the user and the bot linked to your LINE Login channel.
-    ///
-    /// - Parameters:
-    ///   - queue: The callback queue that is used for `completion`. The default value is
-    ///            `.currentMainOrAsync`. For more information, see `CallbackQueue`.
-    ///   - completion: The completion closure to be invoked when the friendship status is returned.
-    /// - Note: The `.profile` permission is required.
-    ///
-    public static func getBotFriendshipStatus(
-        callbackQueue queue: CallbackQueue = .currentMainOrAsync,
-        completionHandler completion: @escaping (Result<GetBotFriendshipStatusRequest.Response, LineSDKError>) -> Void)
-    {
-        let request = GetBotFriendshipStatusRequest()
         Session.shared.send(request, callbackQueue: queue, completionHandler: completion)
     }
 }
