+++ LineSDK/LineSDK/Networking/API/API+Internal.swift
